refactors db code into more files
related #11
