+++ test/server/database/test_helpers.js
@@ -1,0 +1,17 @@
+const makeMockQuery = () => {
+  const queriesMade = []
+  return {
+    queriesMade,
+    query: (queryStr, queryArgs, cb) => {
+      queriesMade.push({
+        text: queryStr,
+        args: queryArgs
+      })
+      cb()
+    }
+  }
+}
+
+module.exports = {
+  makeMockQuery
+}
