+++ test/server/index.js
@@ -1,3 +1,4 @@
 require('./routes.js')
-require('./db.js')
+require('./database/helpers.js')
+require('./database/queries.js')
 require('./image.js')
